refactor(di): split di.go into focused provider files (#768)
internal/di/di.go had grown into a 1050-line wiring file. Splits it
into focused siblings without changing behaviour:

- injector.go   NewInjector, top-level registration order, WireAuditMetrics
- store.go      database + role aliases + metrics deps + instance ID
- backend.go    S3 backends, breaker registry, BackendManager, and the
                optional providers feeding it (encryption, Redis counters,
                object cache)
- workers.go    background worker providers
- lifecycle.go  lifecycle manager and per-mode service registration
- transport.go  S3, admin, UI, notifier, rate limiter, login throttle
- optional.go   invokeOptional and resolveOptional* helpers

services.go (the lifecycle Runner wrappers) is unchanged, as are the
test files. Provider names, signatures, and the registration order in
NewInjector are untouched.

Updates docs/style-guide.md so the Wiring Point section references the
package and the new file layout instead of the old single-file shape.
