+++ .version
@@ -1,1 +1,1 @@
-v0.46.24
+v0.46.25
